requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.4
django-cors-headers>=4.6
django-environ>=0.12
psycopg[binary]>=3.2
gunicorn>=23
whitenoise>=6.8
